Rename src/transformers to src/application Since it is a part of the Application Protocol
