+++ test/transformers/definitions.js
@@ -2,9 +2,9 @@ const { describe, it } = require('mocha');
 const assert = require('assert');
 const { fromPairs, isArray, isObject, isEmpty } = require('lodash');
 
-const compile = require('../../src/transformers/compile');
-const { onExist, on, onWhen, spreadObj } = require('../../src/transformers/helpers');
-const all = require('../../src/transformers/definitions');
+const compile = require('../../src/application/compile');
+const { onExist, on, onWhen, spreadObj } = require('../../src/application/helpers');
+const all = require('../../src/application/definitions');
 
 const transformers = fromPairs(all.map(t => [t.cmd, t]));
 
@@ -59,7 +59,7 @@ const test = (tInput, tOutput, tMock) => {
     }
 };
 
-describe('transformers', () => {
+describe('definitions', () => {
     it('should transform DANMU_MSG correctly', () => {
         const inputTemplate = {
             info: [
